terminology fixes
