+++ kie-dmn/kie-dmn-validation/src/test/java/org/kie/dmn/validation/dtanalysis/AbstractDTAnalysisTest.java
@@ -25,7 +25,6 @@ import org.drools.javaparser.printer.PrettyPrinterConfiguration;
 import org.kie.dmn.api.core.DMNMessage;
 import org.kie.dmn.validation.AbstractValidatorTest;
 import org.kie.dmn.validation.ValidatorUtil;
-import org.kie.dmn.validation.dtanalysis.model.DMNDTAnalysisMessage;
 import org.kie.dmn.validation.dtanalysis.model.DTAnalysis;
 import org.kie.dmn.validation.dtanalysis.model.Hyperrectangle;
 import org.kie.dmn.validation.dtanalysis.model.Overlap;
